Linux Plugin Binary for X-Plane 12
This is the Linux only version plugin to be installed in the X-Plane\Resources\plugins directory. This plugin was compiled against the X-Plane SDK 401  This is the 32-bit version & the 64-bit version is in the 64 sub-directory. Note X-Plane only runs in 64-bit mode as of X-Plane 11. The 32-bit version is not really used anymore